+++ .envs/.local/.postgres
@@ -2,6 +2,6 @@
 # ------------------------------------------------------------------------------
 POSTGRES_HOST=postgres
 POSTGRES_PORT=5432
-POSTGRES_DB=braavos
+POSTGRES_DB=tutorial
 POSTGRES_USER=django
 POSTGRES_PASSWORD=p@ssword
