+++ config/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for braavos project.
+Django settings for tutorial project.
 
 Generated by 'django-admin startproject' using Django 5.2.
 
@@ -42,7 +42,7 @@ INSTALLED_APPS = [
     "django.contrib.messages",
     "django.contrib.staticfiles",
     "rest_framework",
-    "braavos.quickstart",
+    "tutorial.quickstart",
 ]
 
 MIDDLEWARE = [
