+++ config/urls.py
@@ -1,5 +1,5 @@
 """
-URL configuration for braavos project.
+URL configuration for tutorial project.
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/5.2/topics/http/urls/
@@ -19,6 +19,6 @@ from django.contrib import admin
 from django.urls import include, path
 
 urlpatterns = [
-    path("", include("braavos.quickstart.urls")),
+    path("", include("tutorial.quickstart.urls")),
     path("admin/", admin.site.urls),
 ]
