+++ tutorial/quickstart/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class QuickstartConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "braavos.quickstart"
+    name = "tutorial.quickstart"
